+++ Changelog.md
@@ -1,3 +1,7 @@
+# Development
+
+* Fix stack trace on parse error
+
 # 1.0.1 - 2022-03-11
 
 * Update gems to support Ruby 3.0.  Now requires >= Ruby 2.5.